requirements.txt
# FinChain Intelligence Network (FIN) - Dependencies

# Core dependencies
langchain>=0.0.347
crewai==0.85.0
# Fetch.ai uAgents package
uagents>=0.6.0

# Web Interface
gradio>=3.50.0
plotly>=5.18.0
matplotlib>=3.8.2
pandas>=2.2.0

# Machine Learning
tensorflow>=2.15.0
torch>=2.2.1
scikit-learn>=1.4.0
numpy>=1.26.3

# Blockchain & Crypto
web3>=6.15.1
# eth-brownie has been updated to brownie
eth-brownie>=1.19.3
# solana SDK might need a different name
solana>=0.30.0

# Natural Language Processing
transformers>=4.38.0
sentence-transformers>=2.2.2

# API Integration
requests>=2.31.0
aiohttp>=3.9.1

# Configuration
pyyaml>=6.0.1

# Development tools
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
flake8>=7.0.0
jupyter>=1.0.0